+++ Module.php
@@ -56,26 +56,6 @@ class Module extends \humhub\components\Module
         return $maxNumber;
     }
 
-    public function getUserIcon()
-    {
-        $icon = $this->settings->get('icon');
-        $color = $this->settings->get('color');
-
-		    $tooltip_message = Yii::t('VerifiedModule.base', 'Verified User');
-		
-		    return Icon::get($icon, ['color' => $color, 'tooltip' => $tooltip_message]);
-    }
-	
-    public function getSpaceIcon()
-    {
-        $icon = $this->settings->get('icon');
-        $color = $this->settings->get('color');
-
-		    $tooltip_message = Yii::t('VerifiedModule.base', 'Verified Space');
-
-		    return Icon::get($icon, ['color' => $color, 'tooltip' => $tooltip_message]);
-    }
-
     /**
      * @inheritdoc
      */